chelsio/chtls: bug 40301: fix panic during unload reload chtls
While offloading the connection to tls-toe,
"@sk_prot_creator: sk_prot of original sock creator"
was not preserved, causing a hole when it is returned
back to infra(stack) as Time-wait socket connection.
Due to which when the tw_timer_handler gets executed
to kill the connection, we see a PANIC in inet_twsk_free.

Signed-off-by: Vinay Kumar Yadav <udai.sharma@chelsio.com>
Signed-off-by: Vinay Kumar Yadav <vinay.yadav@chelsio.com>